create thread.c and .h
