+++ Makefile
@@ -13,15 +13,15 @@ OBJECTS := $(SRC:src/%.c=$(OBJ_DIR)/%.o)
 all: build $(EXEC_DIR)/$(TARGET)
 
 $(OBJ_DIR)/%.o: src/%.c
-	@mkdir -p $(@D) # Aqui corrigimos com tab
+	@mkdir -p $(@D) 
 	$(CC) $(CXXFLAGS) $(INCLUDE) -c $< -o $@
 
 $(EXEC_DIR)/$(TARGET): $(OBJECTS)
-	@mkdir -p $(@D) # Aqui corrigimos com tab
+	@mkdir -p $(@D) 
 	$(CC) $(OBJECTS) $(CXXFLAGS) $(LDFLAGS) -o $@
 
 build:
-	@mkdir -p $(EXEC_DIR) # Aqui corrigimos com tab
+	@mkdir -p $(EXEC_DIR) 
 	@mkdir -p $(OBJ_DIR)
 
 debug: CXXFLAGS += -DDEBUG -g
