+++ src/libs.h
@@ -6,5 +6,6 @@
 #include <string.h>
 #include <stdbool.h>
 #include <ctype.h>
+#include <pthread.h>
 
 #endif
